Create Dockerfile, start work on program bootstrap
A new bootstrapper replaces the old method of running Gunicorn manually.
Some of the initial application parameters are adjustable via
environment variables, this is done to make the system more easily
configurable via a Dockerfile.

A first-time setup server is built into the bootstrapper, so as to make
initial configuration easier. The first-time setup server is configured
to automatically terminate when initial configuration is done, and the
production server is launched immediately following the serialization of
the provided config options.

Along with the bootstrapper, the framework has been laid for more
flexible configuration options. Work is starting on allowing the config
to be stored in a database or redis server, or to be retrieved via and
API endpoint.

Finally, the `server` module has been created. This module is intended
to hold much of the boilerplate code for the app, such that the primary
app functionality stands alone in the root directory. This is done in an
effort to make contributing easier.