QoL and consistency changes
The run script now uses the bootstrapper as opposed to running the app
separately. This means first time setup will run in a dev environment.

The first time setup runner now has a timeout of 5 seconds to wait for
the production server to launch, and displays a loading animation while
pinging the server for a response. The page will also display an error
if the production server fails to launch, instead of redirecting to a
nonexistent web page.

The production server will now display a more useful error message if
the API URL specified does not point to a valid Augur instance. Also,
it will try an extra variation of the API URL just in case before giving
up.

Documentation for server installation now also refers to installing on a
server using Docker, and the documentation for setting up a proxy has
been separated from the systemd setup.

Logging is no longer initialized in the init startup script, but
instead in the utility script.

ServerThread is now wait/terminate compatible with subprocesses.